+++ addOns/selenium/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update minimum ZAP version to 2.16.0.
 - Update Selenium to version 4.27.0.
+- Use WebDriver BiDi with Firefox.
 
 ## [15.30.0] - 2024-09-24
 ### Changed